// build-test/src/types/interface.ts
export interface WorkItemForm {
  page_index?: number;
  page_size?: number;
  status?: null | number[];
  type?: null | number | [];
  times?: string[];
  end_time: string;
  start_time: string;
  sort_name?: string;
  sort_by?: number;
  name: string;
  product_id?: null | number;
  is_drawer?: number | undefined;
}

export interface ProductDynamic {
  page_index?: number;
  page_size?: number;
  type?: null | number;
  create_time?: string[];
  sort_name?: string;
  sort_by?: number;
  product_id?: null | number;
  create_by?: string[];
  create_by_no?: string[];
}